+++ registry/lib/pages/card_page.dart
@@ -1,4 +1,3 @@
-import 'package:any_syntax_highlighter/any_syntax_highlighter.dart';
 import 'package:fcnui_base/fcnui_base.dart';
 import 'package:flutter/material.dart';
 import 'package:registry/pages/page_impl.dart';
@@ -6,7 +5,6 @@ import 'package:registry/ui/default_components/button.dart';
 import 'package:registry/ui/default_components/card.dart';
 import 'package:registry/ui/default_components/form.dart';
 import 'package:registry/ui/default_components/save_button.dart';
-import 'package:registry/ui/layout/default_layout.dart';
 import '../ui/default_components/input.dart';
 import '../ui/default_components/with_label.dart';