+++ registry/lib/pages/pages.dart
@@ -3,3 +3,4 @@ export 'button_page.dart';
 export 'card_page.dart';
 export 'input_page.dart';
 export 'checkbox_page.dart';
+export 'dropdown_page.dart';