+++ fcnui-website/docs/CHANGELOG.md
@@ -8,10 +8,14 @@ Full changelog of the project.
 - Base - `fcnui_base` package which is the core of `fcnui` package.
 - API - `fcnui` package API documentation which provides `json` format components.
 
+## 21.02.2024
+
+- [Registry] Add `dropdown` component.
+- [Homepage] Add `dropdown` documentation.
+
 ## 20.02.2024
 
 - Add TextEditingController parameter for `Input` component.
-- 
 
 ## 19.02.2024